+++ client/src/Views/SignIn.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const SignIn = () => {
+  return (
+    <div>Sign In</div>
+  )
+}
+ 
+export default SignIn;
