+++ client/src/Views/SignUp.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const SignUp = () => {
+  return (
+    <div>Sign Up</div>
+  )
+}
+
+export default SignUp;
