+++ client/src/Views/Users.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Users = () => {
+  return (
+    <div>Users</div>
+  )
+}
+
+export default Users;
