+++ client/src/components/App/App.js
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+import { BrowserRouter as Router } from 'react-router-dom';
+import Routes from '../Routes/Routes';
+import logo from './logo.svg';
+import './App.css';
+
+class App extends Component {
+  render() {
+    return (
+      <Router>
+        <div className="App">
+          <header className="App-header">
+            <img src={logo} className="App-logo" alt="logo" />
+            <h1 className="App-title">Authentication using JWTs</h1>
+          </header>
+          <Routes />
+        </div>
+      </Router>
+    );
+  }
+}
+
+export default App;
